Fix network error handling and extract shared API service logic
- Fix recordUsage, deleteProgram, and deleteTemplate methods to wrap
  URLError in RepositoryError types for consistent error handling
- Extract duplicated helper methods (addAuthHeader, performRequest,
  validateResponse) into new BaseAPIService base class
- Reduce code duplication by ~313 lines across 5 API service classes
- Update CLAUDE.md with testing and change summary requirements
