+++ Nippardation/Services/API/SyncAPIService.swift
@@ -7,14 +7,10 @@
 
 import Foundation
 
-final class SyncAPIService: SyncAPIServiceProtocol, @unchecked Sendable {
+final class SyncAPIService: BaseAPIService, SyncAPIServiceProtocol, @unchecked Sendable {
 
-    private let session: URLSession
-    private let decoder = JSONDecoder.apiDecoder
-    private let encoder = JSONEncoder.apiEncoder
-
-    init(session: URLSession = .shared) {
-        self.session = session
+    override init(session: URLSession = .shared) {
+        super.init(session: session)
     }
 
     // MARK: - SyncAPIServiceProtocol
@@ -33,78 +29,18 @@ final class SyncAPIService: SyncAPIServiceProtocol, @unchecked Sendable {
         return try await performRequest(request)
     }
 
-    // MARK: - Private Helpers
+    // MARK: - Override for Sync-specific response handling
 
-    private func addAuthHeader(to request: inout URLRequest) async throws {
-        guard let token = try await AuthManager.shared.getIDToken() else {
-            throw RepositoryError.unauthorized
-        }
-        request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-    }
-
-    private func performRequest<T: Decodable>(_ request: URLRequest) async throws -> T {
-        let data: Data
-        let response: URLResponse
-
-        do {
-            (data, response) = try await session.data(for: request)
-        } catch let error as URLError {
-            switch error.code {
-            case .notConnectedToInternet, .networkConnectionLost:
-                throw RepositoryError.networkUnavailable
-            case .timedOut:
-                throw RepositoryError.timeout
-            default:
-                throw RepositoryError.unknown(error)
-            }
-        }
-
-        try validateResponse(response, data: data)
-
-        do {
-            return try decoder.decode(T.self, from: data)
-        } catch {
-            #if DEBUG
-            print("Decoding error: \(error)")
-            if let json = String(data: data, encoding: .utf8) {
-                print("Response: \(json)")
-            }
-            #endif
-            throw RepositoryError.unknown(error)
-        }
-    }
-
-    private func validateResponse(_ response: URLResponse, data: Data? = nil) throws {
+    override func validateResponse(_ response: URLResponse, data: Data? = nil) throws {
         guard let httpResponse = response as? HTTPURLResponse else {
             throw RepositoryError.unknown(nil)
         }
 
-        switch httpResponse.statusCode {
-        case 200...299:
+        // 409 Conflict is valid for sync - let caller handle via response data
+        if httpResponse.statusCode == 409 {
             return
-        case 401:
-            throw RepositoryError.unauthorized
-        case 404:
-            throw RepositoryError.notFound
-        case 409:
-            // Sync conflict - let caller handle via response data
-            return
-        case 422:
-            let message = data.flatMap { String(data: $0, encoding: .utf8) } ?? "Validation failed"
-            throw RepositoryError.validationError(message)
-        case 429:
-            let retryAfter = httpResponse.value(forHTTPHeaderField: "Retry-After")
-                .flatMap { Double($0) }
-            throw RepositoryError.rateLimited(retryAfter: retryAfter)
-        case 400...499:
-            let message = data.flatMap { String(data: $0, encoding: .utf8) }
-            throw RepositoryError.serverError(statusCode: httpResponse.statusCode, message: message)
-        case 500...599:
-            let message = data.flatMap { String(data: $0, encoding: .utf8) }
-            throw RepositoryError.serverError(statusCode: httpResponse.statusCode, message: message)
-        default:
-            throw RepositoryError.unknown(nil)
         }
+
+        try super.validateResponse(response, data: data)
     }
 }
